+++ src/l1j/server/server/serverpackets/S_OwnCharPack.java
@@ -33,14 +33,14 @@ public class S_OwnCharPack extends ServerBasePacket {
 	private static Logger _log = Logger
 			.getLogger(S_OwnCharPack.class.getName());
 
-	private static final int STATUS_POISON = 1;
-	private static final int STATUS_INVISIBLE = 2;
-	private static final int STATUS_PC = 4;
-	private static final int STATUS_FREEZE = 8;
-	private static final int STATUS_BRAVE = 16;
-	private static final int STATUS_ELFBRAVE = 32;
-	private static final int STATUS_FASTMOVABLE = 64;
-	private static final int STATUS_GHOST = 128;
+	private static final short STATUS_POISON = 1;
+	private static final short STATUS_INVISIBLE = 2;
+	private static final short STATUS_PC = 4;
+	private static final short STATUS_FREEZE = 8;
+	private static final short STATUS_BRAVE = 16;
+	private static final short STATUS_ELFBRAVE = 32;
+	private static final short STATUS_FASTMOVABLE = 64;
+	private static final short STATUS_GHOST = 128;
 
 	private byte[] _byte = null;
 
@@ -49,7 +49,7 @@ public class S_OwnCharPack extends ServerBasePacket {
 	}
 
 	private void buildPacket(L1PcInstance pc) {
-		int status = STATUS_PC;
+		short status = STATUS_PC;
 
 		// グール毒みたいな綠の毒
 		// if (pc.isPoison()) {
